ambient light and camera parsing are working

+++ libs/libft/libft.h
@@ -46,7 +46,7 @@ char		*ft_strtrim(char const *s1, char const *set);
 char		**ft_split(char const *s, char c);
 char		*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 void		ft_striteri(char *s, void (*f)(unsigned int, char*));
-long int	ft_atoi(const char *str);
+int			ft_atoi(const char *str);
 int			ft_isalpha(int c);
 int			ft_toupper(int c);
 int			ft_isdigit(int c);
